Start moving components out of jazzbar
