+++ src/components/attendee/MediaPlayer/index.tsx
@@ -1,0 +1,1 @@
+export { MediaPlayer } from "./MediaPlayer";
